+++ src/mjlab/actuator/__init__.py
@@ -5,6 +5,12 @@ from mjlab.actuator.actuator import ActuatorCfg as ActuatorCfg
 from mjlab.actuator.actuator import ActuatorCmd as ActuatorCmd
 from mjlab.actuator.actuator import CommandField as CommandField
 from mjlab.actuator.builtin_actuator import (
+  BuiltinDcMotorActuator as BuiltinDcMotorActuator,
+)
+from mjlab.actuator.builtin_actuator import (
+  BuiltinDcMotorActuatorCfg as BuiltinDcMotorActuatorCfg,
+)
+from mjlab.actuator.builtin_actuator import (
   BuiltinMotorActuator as BuiltinMotorActuator,
 )
 from mjlab.actuator.builtin_actuator import (